pyproject.toml
[tool.poetry]
name = "selenium-pixels"
version = "0.1.0"
description = ""
authors = ["Lorenzo Delgado <lorenzodelgado.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.3.0"
pillow = "^7.2.0"
selenium = "^3.141.0"
pytest = "^5.4.3"
pytesseract = "^0.3.4"

[tool.poetry.dev-dependencies]
isort = "^5.0.9"
typing = "^3.7.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
